Implement new IMU and do some code cleanup

+++ Fin_NRC_BB_code/Filter.cpp
@@ -24,7 +24,7 @@ Filter::Filter(unsigned size)
   this->bufferSize = size; //"this" unnecessary
 
   //allocate our array
-  buffer = new int[size];
+  buffer = new double[size];
 
   //reset the values in the buffer
   reset();
